:white_check_mark: Add testing configuration
