+++ jest.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+  coveragePathIgnorePatterns: ['/node_modules/'],
+  globals: {
+    'ts-jest': {
+      tsConfig: 'tsconfig.json',
+    },
+  },
+  moduleFileExtensions: ['ts', 'js', 'json'],
+  transform: {
+    '^.+\\.(ts|tsx)$': 'ts-jest',
+  },
+  testPathIgnorePatterns: ['/node_modules/'],
+  testMatch: ['**/?(*.)+(spec|test).ts?(x)'],
+  testEnvironment: 'node',
+};
