+++ package.json
@@ -6,19 +6,24 @@
   "author": "emiling <scopradical@gmail.com>",
   "license": "MIT",
   "scripts": {
-    "start": "tsc && node ./dist"
+    "start": "tsc && node ./dist",
+    "test": "jest"
   },
   "devDependencies": {
+    "@types/jest": "^25.2.1",
     "@types/node": "^13.11.1",
     "@typescript-eslint/eslint-plugin": "^2.26.0",
     "@typescript-eslint/parser": "^2.26.0",
     "eslint": "^6.8.0",
     "eslint-config-google": "^0.14.0",
     "eslint-config-prettier": "^6.10.1",
+    "eslint-plugin-jest": "^23.8.2",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^3.1.2",
     "gts": "^1.1.2",
+    "jest": "^25.3.0",
     "prettier": "^2.0.4",
+    "ts-jest": "^25.3.1",
     "typescript": "^3.8.3"
   }
 }
