+++ src/Calculator.ts
@@ -1,0 +1,8 @@
+export class Calculator {
+  public add(a: number, b: number) {
+    return a + b;
+  }
+  private doNotUseSubtract(a: number, b: number) {
+    return a - b;
+  }
+}
