+++ tsconfig.json
@@ -49,7 +49,7 @@
     // "typeRoots": [],                       /* List of folders to include type definitions from. */
     "types": [
       "node",
-      // "jest"
+      "jest"
     ],                           /* Type declaration files to be included in compilation. */
     // "allowSyntheticDefaultImports": true,  /* Allow default imports from modules with no default export. This does not affect code emit, just typechecking. */
     // "esModuleInterop": true,                  /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
